SL-1472: Refactor
